Simple Example of Class and Object
In this Example We have created a ExampleOfClassAndobject Class. and
That have three instance variable, (Instance variable: A variable that
is created inside of class But outside the method. it is called Instance
variable.)
it is also called data member of class. there are  id, name, marks, Id
is int type variable.name is String type variable. and marks is double
type variable. After that we have used main method of class. and we have
created the object of the ExampleOfClassAndobject by new keyword(The new
keyword is used to allocate memory at runtime) and printing the object
value. In this Program values are not printed becouse values are not
initized. so it is printed null or zero . there are by default values of
int,String and double.