Simple Example of class and object with initilized values
In this Example We have created a ExampleOfObjectAndclass (Class). and
That have three instance variable, (Instance variable: A variable that
is created inside of class But outside the method. it is called Instance
variable.)
it is also called data member of class. there are  id, name, marks, Id
is int type variable.name is String type variable. and marks is double
type variable.
and that have two method. one is parameterize and second is simple
method. After that we have used main method of class. and we have
created the object of the ExampleOfObjectAndclass(class) by new
keyword(The new keyword is used to allocate memory at runtime) here
creating two object of the class. and printing the object value. In this
Program values are  printed becouse values are initized. so these are
printed Rolll number,Student name, marks . you have one think. in this
program. i took a method parametrized. how value is print. you see.
object is an instance variable of class.
object gets the memory in heap area and reference variable refers to the
object allocated in heap memory area. Here example1 and example2 both
are reference variable that refers to the object allocated in memory.